refactor(session): as jsdoc comments
